+++ requirements.txt
@@ -5,3 +5,4 @@ email-validator>=1.1.3
 psycopg2-binary>=2.9.1
 redis>=3.5.3
 python-multipart>=0.0.5
+python-dateutil==2.9.0
